コードの型安全性と堅牢性の向上、およびTODOの追加
- `Line.cpp` の `Line::CreateVertexBuffer()` で `vertexBufferView_.SizeInBytes` の型変換を明示的に `static_cast<UINT>` に修正。型の安全性を向上。
- `LineManager.cpp` の `LineManager::ClearLines()` に `line_->DrawLine` 呼び出しを追加。「TODO: 問題あり」のコメントを追記し、線の太さサポートの必要性を明記。
- `LineManager.cpp` の `LineManager::DrawLightRays` で `alpha2` の計算式を削除。
- `main.cpp` の `WinMain` で引数の妥当性を確認するコードを追加。不正な場合はエラー終了する処理を実装し、プログラムの堅牢性を向上。

+++ engine/3d/Line.cpp
@@ -151,7 +151,7 @@ void Line::CreateVertexBuffer() {
 	// バーテックスバッファビューの設定
     vertexBufferView_.BufferLocation = vertexBuffer_->GetGPUVirtualAddress();
 	// バイトサイズ
-    vertexBufferView_.SizeInBytes = bufferSize;
+	vertexBufferView_.SizeInBytes = static_cast<UINT>( bufferSize );
 	// ストライド
     vertexBufferView_.StrideInBytes = sizeof(LineVertex);
 }
